Develop

+++ Reads/Controllers/BooksController.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
 using Reads.Models;
+using Reads.Validators;
 
 namespace Reads.Controllers
 {
@@ -9,10 +10,12 @@ namespace Reads.Controllers
     public class BooksController : ControllerBase
     {
         private readonly IBookRepository _bookRepository;
+        private readonly BookValidator _validator;
 
-        public BooksController(IBookRepository bookRepository)
+        public BooksController(IBookRepository bookRepository, BookValidator validator)
         {
             _bookRepository = bookRepository;
+            _validator = validator;
         }
 
         // GET api/books
@@ -20,16 +23,17 @@ namespace Reads.Controllers
         public ActionResult<IEnumerable<Book>> Get()
         {
             List<Book> books = _bookRepository.GetAll().Result;
-            return Ok(books);
+            return books;
         }
 
         // POST api/books
         [HttpPost]
         public ActionResult Post([FromBody] Book book)
         {
-            if (!ModelState.IsValid)
+            var validation = _validator.Validate(book);
+            if (!validation.IsValid)
             {
-                return BadRequest(ModelState);
+                return BadRequest(validation.ToString());
             }
             _bookRepository.Add(book);
             return Ok();
@@ -39,9 +43,10 @@ namespace Reads.Controllers
         [HttpPut]
         public ActionResult Put([FromBody] Book book)
         {
-            if (!ModelState.IsValid)
+            var validation = _validator.Validate(book);
+            if (!validation.IsValid)
             {
-                return BadRequest(ModelState);
+                return BadRequest(validation.ToString());
             }
             _bookRepository.Update(book);
             return Ok();
@@ -56,6 +61,10 @@ namespace Reads.Controllers
                 return BadRequest(ModelState);
             }
             var book = _bookRepository.Get(id).Result;
+            if (book == null)
+            {
+                return BadRequest($"No book found for id: {id}");
+            }
             _bookRepository.Delete(book);
             return Ok();
         }
