+++ Reads/Controllers/CategoriesController.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
-using Reads;
 using Reads.Models;
+using Reads.Validators;
 
 namespace Reads.Controllers
 {
@@ -10,27 +10,30 @@ namespace Reads.Controllers
     public class CategoriesController : ControllerBase
     {
         private readonly ICategoryRepository _categoryRepository;
+        private readonly CategoryValidator _validator;
 
-        public CategoriesController(ICategoryRepository categoryRepository)
+        public CategoriesController(ICategoryRepository categoryRepository, CategoryValidator validator)
         {
             _categoryRepository = categoryRepository;
+            _validator = validator;
         }
 
         // GET api/categories
         [HttpGet]
-        public ActionResult<IEnumerable<string>> Get()
+        public ActionResult<IEnumerable<Category>> Get()
         {
             List<Category> categories = _categoryRepository.GetAll().Result;
-            return Ok(categories);
+            return categories;
         }
 
         // POST api/categories
         [HttpPost]
         public ActionResult Post([FromBody] Category category)
         {
-            if (!ModelState.IsValid)
+            var validation = _validator.Validate(category);
+            if (!validation.IsValid)
             {
-                return BadRequest(ModelState);
+                return BadRequest(validation.ToString());
             }
             _categoryRepository.Add(category);
             return Ok();
@@ -40,9 +43,10 @@ namespace Reads.Controllers
         [HttpPut]
         public ActionResult Put([FromBody] Category category)
         {
-            if (!ModelState.IsValid)
+            var validation = _validator.Validate(category);
+            if (!validation.IsValid)
             {
-                return BadRequest(ModelState);
+                return BadRequest(validation.ToString());
             }
             _categoryRepository.Update(category);
             return Ok();
@@ -57,6 +61,10 @@ namespace Reads.Controllers
                 return BadRequest(ModelState);
             }
             var category = _categoryRepository.Get(id).Result;
+            if (category == null)
+            {
+                return BadRequest($"No category found for id: {id}");
+            }
             _categoryRepository.Delete(category);
             return Ok();
         }
